+++ Sources/Exercise.swift
@@ -15,18 +15,6 @@ public class Exercise: NSManagedObject {}
 
 extension Exercise: UserOrdered {}
 
-/// Struct providing archivable reference to the managed object.
-/// Codable for use in NavigationStack, State, and UserActivity.
-/// Identifiable for use in .sheet and .fullScreenCover.
-/// RawRepresentable so that it can be stored in SceneStorage or AppStorage.
-/// Typed to Exercise for use as a navigationDestination.
-// public typealias ExerciseUriRep = UriRep<Exercise>
-// public extension Exercise {
-//    var uriRep: ExerciseUriRep {
-//        UriRep(value: uriRepresentation)
-//    }
-// }
-
 public extension Exercise {
     // NOTE: does NOT save to context
     static func create(_ context: NSManagedObjectContext, userOrder: Int16) -> Exercise {
@@ -40,7 +28,6 @@ public extension Exercise {
         NSManagedObject.get(context, forURIRepresentation: url) as? Exercise
     }
 
-    // @NSManaged public var name: String?
     var wrappedName: String {
         get { name ?? "unknown" }
         set { name = newValue }
@@ -60,12 +47,4 @@ public extension Exercise {
     var isDone: Bool {
         lastCompletedAt != nil
     }
-
-//    var canAdvance: Bool {
-//        !isDone && !atMax
-//    }
-
-//    var atMax: Bool {
-//        intensityMaxValue <= lastIntensity
-//    }
 }
